Modularize code more, add date search
