+++ 06-Eventonica-SQL/helpers/db/Users.js
@@ -1,93 +1,6 @@
 const db = require("./dbConfig");
 const pgp = require("pg-promise")(); // for helpers
 
-// EVENTS
-function getAllEvents(category, date) {
-  let selectQuery = "SELECT * FROM events";
-  if (category) {
-    selectQuery += ` WHERE category = '${category}'`;
-  }
-  return db
-    .any(selectQuery)
-    .then(events => {
-      return { data: { events } };
-    })
-    .catch(error => {
-      console.error(error);
-      return { error };
-    });
-}
-
-function getEventById(id) {
-  return db
-    .any("SELECT * FROM events WHERE id = $1", id)
-    .then(([event]) => {
-      if (!event) return { error: "Event not found" };
-      else return { data: { event } };
-    })
-    .catch(error => {
-      console.error(error);
-      return { error };
-    });
-}
-
-function createNewEvent(eventProperties) {
-  return db
-    .query(
-      "INSERT INTO events(${this:name}) VALUES(${this:csv})",
-      eventProperties
-    )
-    .then(() =>
-      db.any("SELECT * FROM events WHERE id = $1", eventProperties.id)
-    )
-    .then(([event]) => {
-      if (!event) return { error: "User not created" };
-      else return { data: { event } };
-    })
-    .catch(error => {
-      console.error(error);
-      return { error };
-    });
-}
-
-function updateEventById(id, updatesObj) {
-  const condition = pgp.as.format(" WHERE id = $1", id);
-  let updateQuery = pgp.helpers.update(updatesObj, null, "events") + condition;
-  return db
-    .any(updateQuery)
-    .then(() => db.any("SELECT * FROM events WHERE id = $1", id))
-    .then(([event]) => {
-      if (!event) return { error: "Event not found" };
-      else return { data: { event } };
-    })
-    .catch(error => {
-      console.error(error);
-      return { error };
-    });
-}
-
-function deleteEventById(id) {
-  let deletedEvent;
-  return db
-    .any("SELECT * FROM events WHERE id = $1", id)
-    .then(([event]) => {
-      if (!event) return { error: "Event not found" };
-      else {
-        deletedEvent = event;
-        return db.any("DELETE FROM events WHERE id = $1", id);
-      }
-    })
-    .then(res => {
-      if (res.error) return res;
-      else return { data: { event: deletedEvent } };
-    })
-    .catch(error => {
-      console.error(error);
-      return { error };
-    });
-}
-
-// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ USERS ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ //
 function getAllUsers() {
   return db
     .any("SELECT * FROM users")
@@ -177,12 +90,12 @@ async function addPersonalEventToUserById(userId, eventId) {
     const [
       saved
     ] = await db.any(
-      "SELECT * FROM saved_events WHERE user_id = $1 AND event_id = $2;",
+      "SELECT * FROM user_events WHERE user_id = $1 AND event_id = $2;",
       [userId, eventId]
     );
     if (user && event && !saved) {
       return db
-        .any("INSERT INTO saved_events (user_id, event_id) VALUES ($1, $2);", [
+        .any("INSERT INTO user_events (user_id, event_id) VALUES ($1, $2);", [
           userId,
           eventId
         ])
@@ -206,19 +119,10 @@ async function addPersonalEventToUserById(userId, eventId) {
 }
 
 module.exports = {
-  Events: {
-    getAll: getAllEvents,
-    getById: getEventById,
-    createNew: createNewEvent,
-    updateById: updateEventById,
-    deleteById: deleteEventById
-  },
-  Users: {
-    getAll: getAllUsers,
-    getById: getUserById,
-    createNew: createNewUser,
-    updateById: updateUserById,
-    deleteById: deleteUserById,
-    addPersonalEventById: addPersonalEventToUserById
-  }
+  getAll: getAllUsers,
+  getById: getUserById,
+  createNew: createNewUser,
+  updateById: updateUserById,
+  deleteById: deleteUserById,
+  addPersonalEventById: addPersonalEventToUserById
 };
