+++ 06-Eventonica-SQL/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "eventonica",
+  "name": "Eventonica",
   "version": "6.0.0",
   "description": "This project is an event managing web app, built according to the instructions found (here)[https://github.com/Techtonica/curriculum/tree/b6b700cfd7d199ccb952d79b1fe9ab09ac265754/projects/eventonica].",
   "main": "index.js",
